Fix ProjectAccessCodeValidator to accept 10 digit numbers. Fix ProjectLobbyStateController to use service-to-service call to ProjectApi to get a project (user will not be authorized at several points to get the project). Fix null reference exception accessing ProjectGuestContext properties within SetProjectGuestContextAsync, as it will not have been created yet. Switch back to ICache from ICacheAsync since ICache inherits from ICacheAsync.

+++ Synthesis.Guest.Modules/Validators/ProjectAccessCodeValidator.cs
@@ -11,7 +11,7 @@ namespace Synthesis.GuestService.Validators
 
             // TODCO: CU-598: Figure why this is causing a null ref exception
             RuleFor(request => request)
-                .Must(x => int.TryParse(x, out var _))
+                .Must(x => long.TryParse(x, out var _))
                 .WithMessage("The project access code must be a number");
         }
     }
